+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 - Add breadcrumbs and sticky lines support
+- Add support for IntelliJ 2025.
 
 ## [1.7.1] - 2025-01-12