+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add breadcrumbs and sticky lines support
 - Add support for IntelliJ 2025.
 - Add support for optional boolean argument in cc_create
+- Update types and trigger names to match the latest decompiler output
 
 ## [1.7.1] - 2025-01-12