+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add support for optional boolean argument in cc_create
 - Update types and trigger names to match the latest decompiler output
 - Add prefix and postfix expression support
+- Allow the "varp" type to be referenced
 
 ## [1.7.1] - 2025-01-12