+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Update types and trigger names to match the latest decompiler output
 - Add prefix and postfix expression support
 - Allow the "varp" type to be referenced
+- Disallow "+" and "-" to be part of identifiers
 
 ## [1.7.1] - 2025-01-12