+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [1.8.0] - 2025-04-17
+
 - Add breadcrumbs and sticky lines support
 - Add support for IntelliJ 2025.
 - Add support for optional boolean argument in cc_create
@@ -158,7 +160,8 @@
 
 - The new file action no longer shows on top of everything else.
 
-[Unreleased]: https://github.com/waleedyaseen/intellij-runescript/compare/v1.7.1...HEAD
+[Unreleased]: https://github.com/waleedyaseen/intellij-runescript/compare/v1.8.0...HEAD
+[1.8.0]: https://github.com/waleedyaseen/intellij-runescript/compare/v1.7.1...v1.8.0
 [1.7.1]: https://github.com/waleedyaseen/intellij-runescript/compare/v1.7.0...v1.7.1
 [1.7.0]: https://github.com/waleedyaseen/intellij-runescript/compare/v1.6.0...v1.7.0
 [1.6.0]: https://github.com/waleedyaseen/intellij-runescript/compare/v1.5.1...v1.6.0